r9009: directory not empty is not an error on failure to delete directory in delete on close (This used to be commit dada509f5e374872cded9035611c9d4fd9f4c4c7)

+++ source4/ntvfs/posix/pvfs_open.c
@@ -75,7 +75,7 @@ static int pvfs_dir_handle_destructor(void *p)
 			DEBUG(0,("Warning: xattr rmdir hook failed for '%s' - %s\n",
 				 h->name->full_name, nt_errstr(status)));
 		}
-		if (rmdir(h->name->full_name) != 0) {
+		if (rmdir(h->name->full_name) != 0 && errno != ENOTEMPTY) {
 			DEBUG(0,("pvfs_close: failed to rmdir '%s' - %s\n", 
 				 h->name->full_name, strerror(errno)));
 		}
